docs: fix kotlin samples not showing
Close: #214
