--- test-suite-kotlin/src/test/kotlin/io/micronaut/docs/ioc/validation/custom/HolidayService.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright 2017-2020 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.docs.ioc.validation.custom
-
-import java.time.Duration
-import jakarta.inject.Singleton
-import jakarta.validation.constraints.NotBlank
-
-// tag::class[]
-@Singleton
-open class HolidayService {
-
-    open fun startHoliday(@NotBlank person: String,
-                          @DurationPattern duration: String): String {
-        val d = Duration.parse(duration)
-        return "Person $person is off on holiday for ${d.toMinutes()} minutes"
-    }
-}
-// end::class[]
