+++ test-suite-kotlin/src/test/kotlin/io/micronaut/docs/validation/PersonServiceSpec.kt
@@ -1,26 +1,28 @@
-package io.micronaut.docs.ioc.validation.custom
+package io.micronaut.docs.validation
 
+// tag::imports[]
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
 import jakarta.inject.Inject
 import jakarta.validation.ConstraintViolationException
+// end::imports[]
 
+// tag::test[]
 @MicronautTest
-internal class DurationPatternValidatorSpec {
+class PersonServiceSpec {
 
-    // tag::test[]
     @Inject
-    lateinit var holidayService: HolidayService
+    lateinit var personService: PersonService
 
     @Test
-    fun testCustomValidator() {
+    fun testThatNameIsValidated() {
         val exception = assertThrows(ConstraintViolationException::class.java) {
-            holidayService.startHoliday("Fred", "junk") // <1>
+            personService.sayHello("") // <1>
         }
 
-        assertEquals("startHoliday.duration: invalid duration (junk), additional custom message", exception.message) // <2>
+        assertEquals("sayHello.name: must not be blank", exception.message) // <2>
     }
-    // end::test[]
 }
+// end::test[]
