+++ test-suite-kotlin/src/test/kotlin/io/micronaut/docs/validation/custom/DurationPatternValidatorSpec.kt
@@ -1,28 +1,26 @@
-package io.micronaut.docs.ioc.validation
+package io.micronaut.docs.validation.custom
 
-// tag::imports[]
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
 import jakarta.inject.Inject
 import jakarta.validation.ConstraintViolationException
-// end::imports[]
 
-// tag::test[]
 @MicronautTest
-class PersonServiceSpec {
+internal class DurationPatternValidatorSpec {
 
+    // tag::test[]
     @Inject
-    lateinit var personService: PersonService
+    lateinit var holidayService: HolidayService
 
     @Test
-    fun testThatNameIsValidated() {
+    fun testCustomValidator() {
         val exception = assertThrows(ConstraintViolationException::class.java) {
-            personService.sayHello("") // <1>
+            holidayService.startHoliday("Fred", "junk") // <1>
         }
 
-        assertEquals("sayHello.name: must not be blank", exception.message) // <2>
+        assertEquals("startHoliday.duration: invalid duration (junk), additional custom message", exception.message) // <2>
     }
+    // end::test[]
 }
-// end::test[]
