+++ test-suite-kotlin/src/test/kotlin/io/micronaut/docs/validation/itfce/VehicleSpec.kt
@@ -1,4 +1,4 @@
-package io.micronaut.docs.ioc.validation.itfce
+package io.micronaut.docs.validation.itfce
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.micronaut.context.ApplicationContext
